Extract progression engine module because decision logic should not live inline in app.py
