+++ app/backend/progression_engine.py
@@ -1,0 +1,502 @@
+def get_effective_load_increment(exercise, user_settings):
+    if not isinstance(exercise, dict):
+        exercise = {}
+    if not isinstance(user_settings, dict):
+        user_settings = {}
+
+    load_increment = exercise.get("load_increment", None)
+    try:
+        if load_increment is not None:
+            load_increment = float(load_increment)
+            if load_increment >= 0:
+                return load_increment
+    except Exception:
+        pass
+
+    equipment_type = str(exercise.get("equipment_type", "")).strip()
+    increments = user_settings.get("equipment_increments", {})
+    if equipment_type and isinstance(increments, dict):
+        try:
+            val = increments.get(equipment_type, None)
+            if val is not None:
+                val = float(val)
+                if val >= 0:
+                    return val
+        except Exception:
+            pass
+
+    try:
+        step = exercise.get("progression_step", 0)
+        step = float(step)
+        if step >= 0:
+            return step
+    except Exception:
+        pass
+
+    return 0.0
+
+
+def get_effective_recommended_step(exercise):
+    if not isinstance(exercise, dict):
+        return 0.0
+    try:
+        step = exercise.get("recommended_step", exercise.get("progression_step", 0))
+        step = float(step)
+        if step >= 0:
+            return step
+    except Exception:
+        return 0.0
+    return 0.0
+
+
+def compute_next_possible_load(last_load, effective_load_increment):
+    try:
+        last_load = float(last_load)
+        effective_load_increment = float(effective_load_increment)
+    except Exception:
+        return None
+
+    if effective_load_increment <= 0:
+        return last_load
+
+    return last_load + effective_load_increment
+
+
+def parse_number_from_load(value):
+    x = str(value or "").strip().lower().replace("kg", "").strip()
+    if not x:
+        return None
+    try:
+        return float(x.replace(",", "."))
+    except Exception:
+        return None
+
+
+def parse_top_rep(rep_range):
+    txt = str(rep_range or "").strip()
+    if "-" in txt:
+        try:
+            return int(txt.split("-")[-1])
+        except Exception:
+            return None
+    try:
+        return int(txt)
+    except Exception:
+        return None
+
+
+def parse_seconds_value(value):
+    txt = str(value or "").strip().lower().replace("sek", "").replace("sec", "").strip()
+    if not txt:
+        return None
+    try:
+        return int(float(txt.replace(",", ".")))
+    except Exception:
+        return None
+
+
+def evaluate_equipment_constraint(last_load, recommended_step, effective_load_increment, candidate_for_progression):
+    equipment_constraint = False
+    recommended_next_load = None
+    actual_possible_next_load = None
+    secondary_constraints = []
+
+    if candidate_for_progression and last_load is not None:
+        recommended_next_load = float(last_load) + recommended_step
+        actual_possible_next_load = compute_next_possible_load(float(last_load), effective_load_increment)
+
+        ideal_jump = recommended_next_load - float(last_load)
+        actual_jump = actual_possible_next_load - float(last_load)
+
+        if actual_jump > ideal_jump:
+            equipment_constraint = True
+            secondary_constraints.append("equipment_constraint")
+
+    return {
+        "equipment_constraint": equipment_constraint,
+        "secondary_constraints": secondary_constraints,
+        "recommended_next_load": recommended_next_load,
+        "actual_possible_next_load": actual_possible_next_load,
+    }
+
+
+def format_rep_progression_target(current_target, increment=2):
+    txt = str(current_target or "").strip()
+    if not txt:
+        return current_target
+
+    if "-" in txt:
+        parts = txt.split("-")
+        try:
+            lo = int(parts[0].strip())
+            hi = int(parts[-1].strip())
+            return f"{lo + increment}-{hi + increment}"
+        except Exception:
+            return current_target
+
+    try:
+        val = int(txt)
+        return str(val + increment)
+    except Exception:
+        return current_target
+
+
+def format_time_progression_target(current_target, increment=5):
+    secs = parse_seconds_value(current_target)
+    if secs is None:
+        return current_target
+    return f"{secs + increment} sek"
+
+
+def analyze_session_result_for_progression(result):
+    if not isinstance(result, dict):
+        return {
+            "source": "none",
+            "has_sets": False,
+            "first_set_load": None,
+            "first_set_reps": None,
+            "min_reps": None,
+            "hit_failure": False,
+            "load_drop_detected": False,
+        }
+
+    raw_sets = result.get("sets", [])
+    hit_failure = bool(result.get("hit_failure", False))
+
+    if isinstance(raw_sets, list) and raw_sets:
+        clean_sets = []
+        for x in raw_sets:
+            if not isinstance(x, dict):
+                continue
+            reps_raw = str(x.get("reps", "")).strip()
+            load_raw = x.get("load", "")
+            reps_num = None
+            try:
+                reps_num = int(reps_raw) if reps_raw != "" else None
+            except Exception:
+                reps_num = None
+            load_num = parse_number_from_load(load_raw)
+            clean_sets.append({
+                "reps": reps_num,
+                "load": load_num,
+            })
+
+        first_set = clean_sets[0] if clean_sets else {}
+        first_set_load = first_set.get("load")
+        first_set_reps = first_set.get("reps")
+
+        rep_values = [x["reps"] for x in clean_sets if x.get("reps") is not None]
+        min_reps = min(rep_values) if rep_values else None
+
+        load_values = [x["load"] for x in clean_sets if x.get("load") is not None]
+        load_drop_detected = False
+        if len(load_values) >= 2 and load_values[0] is not None:
+            for later in load_values[1:]:
+                if later is not None and later < load_values[0]:
+                    load_drop_detected = True
+                    break
+
+        return {
+            "source": "session_result",
+            "has_sets": True,
+            "first_set_load": first_set_load,
+            "first_set_reps": first_set_reps,
+            "min_reps": min_reps,
+            "hit_failure": hit_failure,
+            "load_drop_detected": load_drop_detected,
+        }
+
+    fallback_load = parse_number_from_load(result.get("load", ""))
+    fallback_reps = None
+    try:
+        achieved_raw = str(result.get("achieved_reps", "")).strip()
+        fallback_reps = int(achieved_raw) if achieved_raw != "" else None
+    except Exception:
+        fallback_reps = None
+
+    return {
+        "source": "session_result",
+        "has_sets": False,
+        "first_set_load": fallback_load,
+        "first_set_reps": fallback_reps,
+        "min_reps": fallback_reps,
+        "hit_failure": hit_failure,
+        "load_drop_detected": False,
+    }
+
+
+def decide_progression_from_context(exercise_id, ctx):
+    step = ctx["step"]
+    start_weight = ctx["start_weight"]
+    latest_result = ctx["latest_result"]
+    analysis = ctx["analysis"]
+    fatigue_score = ctx["fatigue_score"]
+    last_load = ctx["last_load"]
+    last_entry = ctx["last_entry"]
+    recommended_step = float(ctx.get("recommended_step", step or 0) or 0)
+    effective_load_increment = float(ctx.get("effective_load_increment", step or 0) or 0)
+    exercise = ctx.get("exercise", {}) or {}
+    progression_mode = str(exercise.get("progression_mode", "")).strip() or "double_progression"
+    progression_style = str(exercise.get("progression_style", "")).strip()
+
+    if step is None or step == 0:
+        if progression_mode == "reps_only":
+            target_value = None
+            achieved_value = None
+
+            if latest_result:
+                target_raw = str(latest_result.get("target_reps", "")).strip()
+                achieved_raw = str(latest_result.get("achieved_reps", "")).strip()
+                target_value = parse_top_rep(target_raw)
+                achieved_value = parse_top_rep(achieved_raw)
+
+                if progression_style == "time_then_variant":
+                    target_secs = parse_seconds_value(target_raw)
+                    achieved_secs = parse_seconds_value(achieved_raw)
+
+                    if target_secs is not None and achieved_secs is not None and achieved_secs >= target_secs:
+                        return {
+                            "ok": True,
+                            "exercise": exercise_id,
+                            "next_load": None,
+                            "next_target_reps": format_time_progression_target(target_raw, 5),
+                            "progression_decision": "increase_reps",
+                            "progression_reason": "mål-tid ramt, øg tid næste gang",
+                            "equipment_constraint": False,
+                            "recommended_next_load": None,
+                            "actual_possible_next_load": None,
+                            "source": "bodyweight_time_progression",
+                            "start_weight": start_weight,
+                            "step": step,
+                            "recommended_step": recommended_step,
+                            "effective_load_increment": effective_load_increment
+                        }
+
+                if progression_style in ("reps_then_variant", "") and target_value is not None and achieved_value is not None and achieved_value >= target_value:
+                    return {
+                        "ok": True,
+                        "exercise": exercise_id,
+                        "next_load": None,
+                        "next_target_reps": format_rep_progression_target(target_raw, 2),
+                        "progression_decision": "increase_reps",
+                        "progression_reason": "top af rep-interval ramt, øg reps næste gang",
+                        "equipment_constraint": False,
+                        "recommended_next_load": None,
+                        "actual_possible_next_load": None,
+                        "source": "bodyweight_rep_progression",
+                        "start_weight": start_weight,
+                        "step": step,
+                        "recommended_step": recommended_step,
+                        "effective_load_increment": effective_load_increment
+                    }
+
+            return {
+                "ok": True,
+                "exercise": exercise_id,
+                "next_load": None,
+                "next_target_reps": None,
+                "progression_decision": "no_progression",
+                "progression_reason": "ingen progression for denne øvelse endnu",
+                "equipment_constraint": False,
+                "recommended_next_load": None,
+                "actual_possible_next_load": None,
+                "source": "bodyweight_no_progression",
+                "start_weight": start_weight,
+                "step": step,
+                "recommended_step": recommended_step,
+                "effective_load_increment": effective_load_increment
+            }
+
+        return {
+            "ok": True,
+            "exercise": exercise_id,
+            "next_load": None,
+            "progression_decision": "no_progression",
+            "progression_reason": "ingen progression for denne øvelse",
+            "equipment_constraint": False,
+            "recommended_next_load": None,
+            "actual_possible_next_load": None,
+            "source": "no_progression",
+            "start_weight": start_weight,
+            "step": step,
+            "recommended_step": recommended_step,
+            "effective_load_increment": effective_load_increment
+        }
+
+    if analysis["source"] == "session_result" and latest_result:
+        target_top = parse_top_rep(latest_result.get("target_reps", ""))
+        first_set_load = analysis.get("first_set_load")
+        first_set_reps = analysis.get("first_set_reps")
+        hit_failure = analysis.get("hit_failure", False)
+        load_drop_detected = analysis.get("load_drop_detected", False)
+
+        progression_reason = "progression holdes"
+
+        candidate_for_progression = bool(
+            first_set_load is not None and
+            target_top is not None and
+            first_set_reps is not None and
+            first_set_reps >= target_top
+        )
+
+        constraint_ctx = evaluate_equipment_constraint(
+            last_load=first_set_load,
+            recommended_step=recommended_step,
+            effective_load_increment=effective_load_increment,
+            candidate_for_progression=candidate_for_progression,
+        )
+        equipment_constraint = constraint_ctx["equipment_constraint"]
+        secondary_constraints = constraint_ctx["secondary_constraints"]
+        recommended_next_load = constraint_ctx["recommended_next_load"]
+        actual_possible_next_load = constraint_ctx["actual_possible_next_load"]
+
+        if first_set_load is None:
+            next_load = start_weight
+            decision = "use_start_weight"
+            progression_reason = "ingen historik, bruger startvægt"
+        elif progression_mode == "none":
+            next_load = int(first_set_load)
+            decision = "no_progression"
+            progression_reason = "ingen progression for denne øvelse"
+        elif hit_failure:
+            next_load = int(first_set_load)
+            decision = "hold"
+            progression_reason = "failure registreret"
+        elif load_drop_detected:
+            next_load = int(first_set_load)
+            decision = "hold"
+            progression_reason = "load-drop mellem sæt"
+        elif fatigue_score >= 2:
+            next_load = int(first_set_load)
+            decision = "hold"
+            progression_reason = "muskeltræthed for høj til progression"
+        elif candidate_for_progression:
+            if equipment_constraint:
+                next_load = int(first_set_load)
+                decision = "hold"
+                progression_reason = "næste mulige spring er for stort"
+            else:
+                next_load = int(actual_possible_next_load)
+                decision = "increase"
+                progression_reason = "top af rep-interval ramt"
+        else:
+            next_load = int(first_set_load)
+            decision = "hold"
+            progression_reason = "progression holdes"
+
+        return {
+            "ok": True,
+            "exercise": exercise_id,
+            "source": "session_result",
+            "last_load": int(first_set_load) if first_set_load is not None else None,
+            "next_load": next_load,
+            "step": step,
+            "recommended_step": recommended_step,
+            "effective_load_increment": effective_load_increment,
+            "start_weight": start_weight,
+            "target_top_reps": target_top,
+            "achieved_reps": first_set_reps,
+            "hit_failure": hit_failure,
+            "load_drop_detected": load_drop_detected,
+            "fatigue_score": fatigue_score,
+            "set_analysis": analysis,
+            "progression_decision": decision,
+            "progression_reason": progression_reason,
+            "equipment_constraint": equipment_constraint,
+            "secondary_constraints": secondary_constraints,
+            "recommended_next_load": recommended_next_load,
+            "actual_possible_next_load": actual_possible_next_load
+        }
+
+    if last_entry is None:
+        return {
+            "ok": True,
+            "exercise": exercise_id,
+            "last_load": None,
+            "next_load": start_weight,
+            "step": step,
+            "recommended_step": recommended_step,
+            "effective_load_increment": effective_load_increment,
+            "progression_decision": "use_start_weight",
+            "progression_reason": "ingen historik, bruger startvægt",
+            "equipment_constraint": False,
+            "secondary_constraints": [],
+            "recommended_next_load": None,
+            "actual_possible_next_load": None,
+            "source": "workout_entry",
+            "start_weight": start_weight,
+        }
+
+    achieved = None
+    try:
+        achieved = int(str(last_entry.get("achieved_reps", "")).strip()) if str(last_entry.get("achieved_reps", "")).strip() else None
+    except Exception:
+        achieved = None
+
+    target_top = parse_top_rep(last_entry.get("reps", ""))
+
+    progression_reason = "progression holdes"
+    decision = "hold"
+    next_load = last_load
+
+    candidate_for_progression = bool(
+        progression_mode != "none" and
+        last_load is not None and
+        target_top is not None and
+        achieved is not None and
+        achieved >= target_top
+    )
+
+    constraint_ctx = evaluate_equipment_constraint(
+        last_load=last_load,
+        recommended_step=recommended_step,
+        effective_load_increment=effective_load_increment,
+        candidate_for_progression=candidate_for_progression,
+    )
+    equipment_constraint = constraint_ctx["equipment_constraint"]
+    secondary_constraints = constraint_ctx["secondary_constraints"]
+    recommended_next_load = constraint_ctx["recommended_next_load"]
+    actual_possible_next_load = constraint_ctx["actual_possible_next_load"]
+
+    if last_load is None:
+        next_load = start_weight
+        decision = "use_start_weight"
+        progression_reason = "ingen historik, bruger startvægt"
+    elif progression_mode == "none":
+        next_load = last_load
+        decision = "no_progression"
+        progression_reason = "ingen progression for denne øvelse"
+    elif fatigue_score >= 2:
+        next_load = last_load
+        decision = "hold"
+        progression_reason = "muskeltræthed for høj til progression"
+    elif candidate_for_progression:
+        if equipment_constraint:
+            next_load = last_load
+            decision = "hold"
+            progression_reason = "næste mulige spring er for stort"
+        else:
+            next_load = int(actual_possible_next_load)
+            decision = "increase"
+            progression_reason = "top af rep-interval ramt"
+
+    return {
+        "ok": True,
+        "exercise": exercise_id,
+        "source": "workout_entry",
+        "last_load": last_load,
+        "next_load": next_load,
+        "step": step,
+        "recommended_step": recommended_step,
+        "effective_load_increment": effective_load_increment,
+        "start_weight": start_weight,
+        "target_top_reps": target_top,
+        "achieved_reps": achieved,
+        "fatigue_score": fatigue_score,
+        "progression_decision": decision,
+        "progression_reason": progression_reason,
+        "equipment_constraint": equipment_constraint,
+        "secondary_constraints": secondary_constraints,
+        "recommended_next_load": recommended_next_load,
+        "actual_possible_next_load": actual_possible_next_load
+    }
